# peer+brontide: when decrypting re-use the allocated ciphertext buf & ensure buf pool buf doesn't escape